HEEDLS-347 Create first page of delegate registration form
