--- DigitalLearningSolutions.Web/Controllers/Register.cs
@@ -1,17 +1,0 @@
-﻿namespace DigitalLearningSolutions.Web.Controllers
-{
-    using Microsoft.AspNetCore.Mvc;
-
-    public class Register : Controller
-    {
-        public IActionResult Index()
-        {
-            if (User.Identity.IsAuthenticated)
-            {
-                return RedirectToAction("Index", "Home");
-            }
-
-            return View();
-        }
-    }
-}
