+++ DigitalLearningSolutions.Web/Views/Register/Index.cshtml
@@ -1,10 +1,72 @@
-﻿@{
-  ViewData["Title"] = "Register";
+@using DigitalLearningSolutions.Web.ViewModels.Register
+@model RegisterViewModel
+@{
+  var firstNameError = ViewData.ModelState["FirstName"]?.Errors?.Count > 0;
+  var lastNameError = ViewData.ModelState["LastName"]?.Errors?.Count > 0;
+  var emailError = ViewData.ModelState["Email"]?.Errors?.Count > 0;
+  var errorHasOccurred = firstNameError || lastNameError || emailError;
+  ViewData["Title"] = errorHasOccurred ? "Error: Register": "Register";
+  var firstNameInputErrorClass = firstNameError ? "nhsuk-input--error" : "";
+  var firstNameFormErrorClass = firstNameError ? "nhsuk-form-group--error" : "";
+  var lastNameInputErrorClass = lastNameError ? "nhsuk-input--error" : "";
+  var lastNameFormErrorClass = lastNameError ? "nhsuk-form-group--error" : "";
+  var emailInputErrorClass = emailError ? "nhsuk-input--error" : "";
+  var emailFormErrorClass = emailError ? "nhsuk-form-group--error" : "";
 }
 
 <div class="nhsuk-grid-row">
   <div class="nhsuk-grid-column-full">
-    <h1 id="page-heading" class="nhsuk-heading-xl">Register</h1>
+    <form class="nhsuk-u-margin-bottom-8" method="post" novalidate asp-action="Index">
+
+      @if (errorHasOccurred) {
+        <partial name="_ErrorSummary"/>
+      }
+      <div class="nhsuk-u-margin-bottom-8">
+        <h1 id="page-heading" class="nhsuk-heading-xl">Register</h1>
+        <p class="nhsuk-body-l">
+          Please enter your personal details to start the registration process
+          to access all Digital Learning Solutions services.
+        </p>
+      </div>
+
+      <div class="nhsuk-form-group @firstNameFormErrorClass">
+        <label class="nhsuk-label" for="FirstName">
+          First name
+        </label>
+        @if (firstNameError) {
+          <span class="nhsuk-error-message" id="first-name-error">
+            <span class="nhsuk-u-visually-hidden">Error:</span> @ViewData.ModelState["FirstName"].Errors[0].ErrorMessage
+          </span>
+        }
+        <input class="nhsuk-input nhsuk-u-width-one-half @firstNameInputErrorClass" type="text" id="FirstName" spellcheck="false" autocomplete="given-name" aria-describedby="first-name-error" asp-for="FirstName"/>
+      </div>
+
+      <div class="nhsuk-form-group @lastNameFormErrorClass">
+        <label class="nhsuk-label" for="LastName">
+          Last name
+        </label>
+        @if (lastNameError) {
+          <span class="nhsuk-error-message" id="last-name-error">
+            <span class="nhsuk-u-visually-hidden">Error:</span> @ViewData.ModelState["LastName"].Errors[0].ErrorMessage
+          </span>
+        }
+        <input class="nhsuk-input nhsuk-u-width-one-half @lastNameInputErrorClass" type="text" id="LastName" spellcheck="false" autocomplete="family-name" aria-describedby="last-name-error" asp-for="LastName"/>
+      </div>
+
+      <div class="nhsuk-form-group @emailFormErrorClass">
+        <label class="nhsuk-label" for="Email">
+          Email
+        </label>
+        @if (emailError) {
+          <span class="nhsuk-error-message" id="email-error">
+            <span class="nhsuk-u-visually-hidden">Error:</span> @ViewData.ModelState["Email"].Errors[0].ErrorMessage
+          </span>
+        }
+        <input class="nhsuk-input nhsuk-u-width-one-half @emailInputErrorClass" type="text" id="Email" spellcheck="false" autocomplete="email" aria-describedby="email-error" asp-for="Email"/>
+      </div>
+
+      <button class="nhsuk-button" type="submit">Next</button>
+    </form>
 
     <p class="nhsuk-u-font-size-24">
       Alternatively, if you already have an account, log in:
